[RN] Add UI for raised hand feature
