+++ react/features/filmstrip/components/native/DominantSpeakerIndicator.js
@@ -1,3 +1,5 @@
+// @flow
+
 import React, { Component } from 'react';
 import { View } from 'react-native';
 
@@ -9,7 +11,7 @@ import styles from './styles';
  * Thumbnail badge showing that the participant is the dominant speaker in
  * the conference.
  */
-export default class DominantSpeakerIndicator extends Component {
+export default class DominantSpeakerIndicator extends Component<{}> {
     /**
      * Implements React's {@link Component#render()}.
      *
@@ -17,10 +19,10 @@ export default class DominantSpeakerIndicator extends Component {
      */
     render() {
         return (
-            <View style = { styles.dominantSpeakerIndicatorBackground }>
+            <View style = { styles.indicatorBackground }>
                 <Icon
                     name = 'dominant-speaker'
-                    style = { styles.dominantSpeakerIndicator } />
+                    style = { styles.indicator } />
             </View>
         );
     }
