+++ react/features/filmstrip/components/web/RaisedHandIndicator.js
@@ -1,13 +1,20 @@
 /* @flow */
 
-import React, { Component } from 'react';
+import React from 'react';
+
+import { connect } from '../../../base/redux';
+
+import AbstractRaisedHandIndicator, {
+    type Props as AbstractProps,
+    _mapStateToProps
+} from '../AbstractRaisedHandIndicator';
 
 import BaseIndicator from './BaseIndicator';
 
 /**
  * The type of the React {@code Component} props of {@link RaisedHandIndicator}.
  */
-type Props = {
+type Props = AbstractProps & {
 
     /**
      * The font-size for the icon.
@@ -25,13 +32,17 @@ type Props = {
  *
  * @extends Component
  */
-class RaisedHandIndicator extends Component<Props> {
+class RaisedHandIndicator extends AbstractRaisedHandIndicator<Props> {
     /**
      * Implements React's {@link Component#render()}.
      *
      * @inheritdoc
      */
     render() {
+        if (!this.props._raisedHand) {
+            return null;
+        }
+
         return (
             <BaseIndicator
                 className = 'raisehandindicator indicator show-inline'
@@ -43,4 +54,4 @@ class RaisedHandIndicator extends Component<Props> {
     }
 }
 
-export default RaisedHandIndicator;
+export default connect(_mapStateToProps)(RaisedHandIndicator);
